Refactor RUN to be simpler

+++ lib/hither.bash
@@ -35,10 +35,26 @@ pg-db-exists() {
   pg_dump -s "${1:-undefined}" &>/dev/null
 }
 
+FAIL=true
+MERGE=false
+SAY=false
+CMD=run
 RUN() {
-  local tmp=$(mktemp)
+  [ $# -gt 0 ] && local cmd=("$@") || local cmd=($CMD)
+  if $SAY; then
+    echo ">>> $cmd"
+  fi
   retval=0
-  stdout=$(run 2>$tmp) || retval=$?
-  stderr=$(< $tmp)
-  rm "$tmp"
+  if $MERGE; then
+    stdout=$("${cmd[@]}" 2>&1) || retval=$?
+    stderr=
+  else
+    local tmp=$(mktemp)
+    stdout=$("${cmd[@]}" 2>$tmp) || retval=$?
+    stderr=$(< $tmp)
+    rm "$tmp"
+  fi
+  if $FAIL && [ $retval -ne 0 ]; then
+    die "Command failed:"$'\n'"$stdout$stderr"
+  fi
 }
